requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
opencv-python>=4.9.0
ultralytics>=8.3.0
python-telegram-bot>=21.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
websockets>=12.0
aiofiles>=24.0.0
httpx>=0.27.0
Pillow>=10.0.0
numpy>=1.26.0
# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# Migrations
alembic>=1.12.0
# Reports
reportlab>=4.0.0
# Scheduler & email
APScheduler>=3.10.0
aiosmtplib>=3.0.0
# Dev / test
pytest>=8.0.0
pytest-asyncio>=0.23.0
